requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3

